refactor: break existing top-level import cycles
